backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.6.0
beanie==1.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.9.0
pydantic-settings==2.5.2
httpx==0.27.0
aiofiles==24.1.0
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
tensorflow==2.17.0
scikit-learn==1.5.1
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
email-validator==2.2.0
